requirements.txt
python-telegram-bot==20.7
pytz
rapidfuzz
openpyxl
pandas
requests
